* Started refactoring all writing functions from VisualLog out to VisualLogBuilder
