+++ docs/update_requirements.sh
@@ -1,2 +1,2 @@
 #/bin/bash
-poetry run poetry export -E docubuild -f requirements.txt -o ./docs/requirements.txt --without-hashes
+poetry run poetry export --with docu -f requirements.txt -o ./docs/requirements.txt --without-hashes
