+++ scistag/logstag/visual_livelog/visual_livelog.py
@@ -93,7 +93,7 @@ class VisualLiveLog(VisualLog):
         :return: True if an update should be executed, e.g. by modifying
         widgets and/or handling begin_update manually.
         """
-        return time.time() - self.last_update >= self._refresh_time_s
+        return time.time() - self.last_update >= self.refresh_time_s
 
     def begin_update(self, clear: bool = False,
                      blocking=False) -> SubLogLock:
@@ -121,10 +121,10 @@ class VisualLiveLog(VisualLog):
         :return: True if an updated is allowed
         """
         self.last_update_start = time.time()
-        if time.time() - self.last_update < self._refresh_time_s:
+        if time.time() - self.last_update < self.refresh_time_s:
             if blocking:
-                while time.time() - self.last_update < self._refresh_time_s:
-                    time.sleep(self._refresh_time_s / 10)
+                while time.time() - self.last_update < self.refresh_time_s:
+                    time.sleep(self.refresh_time_s / 10)
             else:
                 return SubLogLock(log=None)
         sll = self.begin_sub_log("liveLog",
@@ -152,7 +152,7 @@ class VisualLiveLog(VisualLog):
         """
         cur_time = time.time()
         super().end_sub_log()
-        if cur_time - self.last_update > self._refresh_time_s:
+        if cur_time - self.last_update > self.refresh_time_s:
             self.last_update = self.last_update_start
             self.write_to_disk()
             self._frames_since_check += 1
