+++ scistag/logstag/visual_log/pyplot_log_context.py
@@ -1,14 +1,26 @@
-# TODO Work in progress
+"""
+Implements the class :class:`PyPlotLogContext` which allows the easy,
+thread-safe integration of matplotlib plots into a VisualLog.
+"""
 
 from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from scistag.logstag.visual_log import VisualLog
+    from scistag.logstag.visual_log.visual_log_builder import VisualLogBuilder
 
 
 class PyPlotLogContext:
-    def __init__(self, target_log: "VisualLog"):
+    """
+    Defines an enclosed, multi-thread safe PyPlot logging context.
+
+    When the context is left it adds the latest figure to the VisualLog
+    it is associated to.
+
+    Is created via VisualLogBuilder.pyplot, see :meth:`VisualLogBuilder.pyplot`
+    """
+
+    def __init__(self, target_log: "VisualLogBuilder"):
         from scistag.plotstag import MPLock
         self.target_log = target_log
         self.mp_lock = MPLock()
