+++ tools/export_requirements/export_requirements.xt
@@ -1,0 +1,2 @@
+#/bin/bash
+poetry run poetry export --with docu -E "full" -f requirements.txt -o ./requirements.txt --without-hashes
